Moved logic of Interpro management into a Service. The controller just calls the new service.
